+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added new configurable setting _First row is header_ for `Data Edit` mode [#1394](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1394)
 - Added new configurable setting _Trim whitespace_ for `Data Edit` mode [#1395](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1395)
 - Added folding for macros usages in the header line [#1401](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1401)
+- Added code completion & reference resolution for User Rights `Type` column [#1402](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1402)
 
 ### `Access Control Lists` enhancements
 - Create Type/attributes references only within `Target` column [#1400](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1400)